% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_json.R
\name{espn_json}
\alias{espn_json}
\title{Gets full ESPN JSON file (Helper Function)}
\usage{
espn_json(gameID)
}
\arguments{
\item{gameID}{ESPN Game ID}
}
\value{
The entire json file for a given gameID on ESPN's website
}
\description{
Gets full ESPN JSON file (Helper Function)
}
\examples{
ID = 401444869
try(espn_json(ID))
}
\author{
Tyson King
}
