% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_softball_pbp.R
\name{espn_softball_pbp}
\alias{espn_softball_pbp}
\title{Get ESPN Play-by-Play data}
\usage{
espn_softball_pbp(gameID)
}
\arguments{
\item{gameID}{ESPN Game ID}
}
\value{
A data frame of every pitch of the given game
}
\description{
Get ESPN Play-by-Play data
}
\examples{
ID = 401444869
try(espn_softball_pbp(ID))
}
\author{
Tyson King
}
