% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_softball_playerbox.R
\name{espn_softball_playerbox}
\alias{espn_softball_playerbox}
\title{Get ESPN Player Box Score for given game}
\usage{
espn_softball_playerbox(gameID)
}
\arguments{
\item{gameID}{ESPN Game ID}
}
\value{
A named list of dataframes, hitting and pitching
}
\description{
Get ESPN Player Box Score for given game
}
\examples{
ID = 401444869
try(espn_softball_playerbox(ID))
}
\author{
Tyson King
}
