% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_softball_teambox.R
\name{espn_softball_teambox}
\alias{espn_softball_teambox}
\title{Get ESPN Team Box Score for given game}
\usage{
espn_softball_teambox(gameID)
}
\arguments{
\item{gameID}{ESPN Game ID}
}
\value{
A named list of dataframes, Batting, Pitching, and Fielding
}
\description{
Get ESPN Team Box Score for given game
}
\examples{
ID = 401444869
try(espn_softball_teambox(ID))
}
\author{
Tyson King
}
