% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cur_season.R
\name{get_cur_season}
\alias{get_cur_season}
\title{Get the current season}
\usage{
get_cur_season()
}
\value{
integer of current season
}
\description{
Get the current season
}
\examples{
try(get_cur_season())
}
