% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_espn_pbp.R
\name{get_espn_pbp}
\alias{get_espn_pbp}
\title{Get ESPN Play-by-Play data}
\usage{
get_espn_pbp(gameID)
}
\arguments{
\item{gameID}{}
}
\value{
A data frame of every pitch of the given game
}
\description{
Get ESPN Play-by-Play data
}
\examples{
ID = 401444869
try(get_espn_pbp(ID))
}
\author{
Tyson King
}
