% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_espn_playerbox.R
\name{get_espn_playerbox}
\alias{get_espn_playerbox}
\title{Get ESPN Player Box Score}
\usage{
get_espn_playerbox(gameID)
}
\arguments{
\item{gameID}{}
}
\value{
A named list of dataframes, hitting and pitching
}
\description{
Get ESPN Player Box Score
}
\examples{
ID = 401444869
try(get_espn_playerbox(ID))
}
\author{
Tyson King
}
