% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_espn_scoreboard.R
\name{get_espn_scoreboard}
\alias{get_espn_scoreboard}
\title{Get ESPN Softball Scoreboard}
\usage{
get_espn_scoreboard(date)
}
\arguments{
\item{date}{YYYY-MM-DD}
}
\value{
data frame with 21 columns of relevant game information
}
\description{
Get ESPN Softball Scoreboard
}
\examples{
try(get_espn_scoreboard("2022-03-05"))
}
\author{
Tyson King
}
