% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_softball_season_scoreboard.R
\name{get_espn_softball_season_scoreboard}
\alias{get_espn_softball_season_scoreboard}
\title{Get ESPN Softball Season Scoreboard}
\usage{
get_espn_softball_season_scoreboard(season)
}
\arguments{
\item{season}{YYYY (2015 - 2022)}
}
\value{
data frame with data from every game in the given season
}
\description{
Disclaimer: Can take a while to run
}
\examples{
try(get_espn_softball_season_scoreboard(2021))
}
\author{
Tyson King
}
