% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_espn_teambox.R
\name{get_espn_teambox}
\alias{get_espn_teambox}
\title{Get ESPN Team Box Score}
\usage{
get_espn_teambox(gameID)
}
\arguments{
\item{gameID}{}
}
\value{
A named list of dataframes, Batting, Pitching, and Fielding
}
\description{
Get ESPN Team Box Score
}
\examples{
ID = 401444869
try(get_espn_teambox(ID))
}
\author{
Tyson King
}
