% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{Gets full ESPN JSON file (Helper Function)}
\usage{
get_json(gameID)
}
\arguments{
\item{gameID}{}
}
\value{
The entire json file for a given gameID on ESPN's website
}
\description{
Gets full ESPN JSON file (Helper Function)
}
\examples{
ID = 401444869
try(get_json(ID))
}
\author{
Tyson King
}
