% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_massey_scoreboard.R
\name{get_massey_scoreboard}
\alias{get_massey_scoreboard}
\title{Get Massey Scoreboard}
\usage{
get_massey_scoreboard(league)
}
\arguments{
\item{league}{D1, D2, or D3}
}
\value{
The final score of all games through the current date
}
\description{
Get Massey Scoreboard
}
\author{
Tyson King
}
