% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_pbp.R
\name{get_ncaa_pbp}
\alias{get_ncaa_pbp}
\title{Get NCAA play-by-play data for an entire season}
\usage{
get_ncaa_pbp(team_id)
}
\arguments{
\item{team_id}{get this from ncaa_teams function}
}
\value{
dataframe of every play from team's season
}
\description{
only has D1 data for now
}
\examples{
try(get_ncaa_pbp(549186))
}
