% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rankings.R
\name{get_rankings}
\alias{get_rankings}
\title{Get NCAA Softball Team Rankings}
\usage{
get_rankings(source)
}
\arguments{
\item{source}{Source options: RPI, D1Softball, USA Today, or Massey}
}
\value{
data frame of rankings, format changes based on source
}
\description{
Get NCAA Softball Team Rankings
}
\examples{
try(get_rankings("RPI"))

}
\author{
Tyson King
}
