% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_stats.R
\name{get_team_stats}
\alias{get_team_stats}
\title{Access NCAA Individual Stats Rankings}
\usage{
get_team_stats()
}
\value{
named list of data frames for the top 50 of each individual stat on the NCAA website
}
\description{
Access NCAA Individual Stats Rankings
}
\examples{
try(get_team_stats())
}
\author{
Tyson King
}
