% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_espn_season_scoreboard.R
\name{load_espn_season_scoreboard}
\alias{load_espn_season_scoreboard}
\title{Get ESPN Softball Scoreboard}
\usage{
load_espn_season_scoreboard(season)
}
\arguments{
\item{season}{YYYY}
}
\value{
data frame with data from every game in the given season
}
\description{
Disclaimer: Can take a while to run
}
\examples{
try(load_espn_season_scoreboard(2020))
}
\author{
Tyson King
}
