% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_naia_softball_pbp.R
\name{load_naia_softball_pbp}
\alias{load_naia_softball_pbp}
\title{Load NAIA PBP}
\usage{
load_naia_softball_pbp(season)
}
\arguments{
\item{season}{YYYY}
}
\value{
dataframe of all play by play data for an NAIA season
}
\description{
Only supports 2023 season for now
}
\examples{
try(load_naia_softball_pbp(2023))
}
