% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_naia_softball_scoreboard.R
\name{load_naia_softball_scoreboard}
\alias{load_naia_softball_scoreboard}
\title{Load NAIA Scoreboard}
\usage{
load_naia_softball_scoreboard(season)
}
\arguments{
\item{season}{YYYY}
}
\value{
dataframe of all games from NAIA website from selected season
}
\description{
Only supports 2023 season for now
}
\examples{
try(load_naia_softball_scoreboard(2023))
}
