% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_scoreboard.R
\name{load_ncaa_scoreboard}
\alias{load_ncaa_scoreboard}
\title{Get NCAA Softball scoreboard for a given day}
\usage{
load_ncaa_scoreboard(date)
}
\arguments{
\item{date}{"YYYY-MM-DD"}
}
\value{
data frame of team names, their scores, and the game ids
}
\description{
Get NCAA Softball scoreboard for a given day
}
\examples{
date = "2022-03-05"
try(load_ncaa_scoreboard(date))
}
\author{
Tyson King
}
