% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_pbp.R
\name{load_ncaa_softball_pbp}
\alias{load_ncaa_softball_pbp}
\title{Load NCAA PBP}
\usage{
load_ncaa_softball_pbp(season, division = "D1")
}
\arguments{
\item{season}{YYYY}

\item{division}{D1, D2, or D3}
}
\value{
dataframe of all play by play data for an NCAA season
}
\description{
Only supports 2023 season for now
}
\examples{
try(load_ncaa_softball_pbp(2023))
}
