% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_playerbox.R
\name{load_ncaa_softball_playerbox}
\alias{load_ncaa_softball_playerbox}
\title{Load NCAA Player Box Scores}
\usage{
load_ncaa_softball_playerbox(season = 2023, category)
}
\arguments{
\item{season}{YYYY}

\item{category}{Hitting or Pitching}
}
\value{
dataframe of all box scores from NCAA website from selected season
}
\description{
Only supports 2023 data for D1 for now
}
\examples{
try(load_ncaa_softball_playerbox(2023))
}
