% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_rosters.R
\name{load_ncaa_softball_rosters}
\alias{load_ncaa_softball_rosters}
\title{Load NCAA Roster}
\usage{
load_ncaa_softball_rosters(season, division = "D1")
}
\arguments{
\item{season}{YYYY format, pass season or vector of seasons 2021-2023}

\item{division}{"D1", "D2", or "D3"}
}
\value{
dataframe of all rosters from given season(s) and division
}
\description{
Supports 2016-2023 for all divisions
}
\examples{
try(load_ncaa_softball_rosters(2022:2023))
}
