% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_scoreboard.R
\name{load_ncaa_softball_scoreboard}
\alias{load_ncaa_softball_scoreboard}
\title{Load NCAA Scoreboard}
\usage{
load_ncaa_softball_scoreboard(season, division = "D1")
}
\arguments{
\item{season}{YYYY}

\item{division}{"D1", "D2", or "D3"}
}
\value{
dataframe of all games from NCAA website from selected season
}
\description{
Supports 2019-2023 for D1 and only 2023 for D2 and D3
}
\examples{
try(load_ncaa_softball_scoreboard(2023))
}
