% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_team_info.R
\name{load_ncaa_softball_team_info}
\alias{load_ncaa_softball_team_info}
\title{Load all-time info for all NCAA softball teams (all divisions)}
\usage{
load_ncaa_softball_team_info()
}
\value{
df with 9 columns of info on each season of each team
}
\description{
Load all-time info for all NCAA softball teams (all divisions)
}
\examples{
try(load_ncaa_softball_team_info())
}
