% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naia_softball_pbp.R
\name{naia_softball_pbp}
\alias{naia_softball_pbp}
\title{Get NCAA play-by-play data for a given game id and game date}
\usage{
naia_softball_pbp(game_id, game_date)
}
\arguments{
\item{game_id}{NAIA Game ID}

\item{game_date}{YYYY-MM-DD format}
}
\value{
dataframe of every play from a given game
}
\description{
only has D1 data for now
}
\examples{
try(naia_softball_pbp("r7z2", "2023-02-05"))
}
