% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naia_softball_playerbox.R
\name{naia_softball_playerbox}
\alias{naia_softball_playerbox}
\title{Get all box scores for a given game id and date}
\usage{
naia_softball_playerbox(game_id, game_date)
}
\arguments{
\item{game_id}{NAIA Game ID}

\item{game_date}{YYYY-MM-DD format}
}
\value{
named list with two elements, "Hitting" and "Pitching"
Each contains a dataframe of the box scores for that game
}
\description{
Get all box scores for a given game id and date
}
\examples{
try(naia_softball_playerbox("r7z2", "2023-02-05"))
}
