% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naia_softball_scoreboard.R
\name{naia_softball_scoreboard}
\alias{naia_softball_scoreboard}
\title{Get all NAIA softball scores for a given day}
\usage{
naia_softball_scoreboard(date)
}
\arguments{
\item{date}{"YYYY-MM-DD"}
}
\value{
data frame of date, team names and their scores
}
\description{
Get all NAIA softball scores for a given day
}
\examples{
date = "2022-03-05"
try(naia_softball_scoreboard(date))
}
\author{
Tyson King
}
