% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naia_softball_season_scoreboard.R
\name{naia_softball_season_scoreboard}
\alias{naia_softball_season_scoreboard}
\title{Get all NAIA softball scores for a given season}
\usage{
naia_softball_season_scoreboard(season)
}
\arguments{
\item{season}{YYYY Only 2023 for now}
}
\value{
data frame of date, team names and their scores
}
\description{
Get all NAIA softball scores for a given season
}
\examples{
season = "2023"
try(naia_softball_season_scoreboard(season))
}
