% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_softball_pbp.R
\name{ncaa_softball_pbp}
\alias{ncaa_softball_pbp}
\title{Get NCAA play-by-play data for an entire season}
\usage{
ncaa_softball_pbp(game_id)
}
\arguments{
\item{game_id}{NCAA Game ID}
}
\value{
dataframe of every play from team's season
}
\description{
only has D1 data for now
}
\examples{
try(ncaa_softball_pbp(2377327))
}
