% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_softball_playerbox.R
\name{ncaa_softball_playerbox}
\alias{ncaa_softball_playerbox}
\title{Get all box scores for a given game id}
\usage{
ncaa_softball_playerbox(game_id)
}
\arguments{
\item{game_id}{NCAA Game ID}
}
\value{
named list with three elements, "Hitting", "Pitching" and "Fielding".
Each contains a dataframe of the box scores for that game
}
\description{
Get all box scores for a given game id
}
\examples{
try(ncaa_softball_playerbox(2377042))
}
