% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_softball_rankings.R
\name{ncaa_softball_rankings}
\alias{ncaa_softball_rankings}
\title{Get NCAA Softball Team Rankings}
\usage{
ncaa_softball_rankings(source)
}
\arguments{
\item{source}{Source options: RPI, D1Softball, USA Today, or Massey}
}
\value{
data frame of rankings, format changes based on source
}
\description{
Get NCAA Softball Team Rankings
}
\examples{
try(ncaa_softball_rankings("RPI"))
}
\author{
Tyson King
}
