% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_softball_rosters.R
\name{ncaa_softball_rosters}
\alias{ncaa_softball_rosters}
\title{Get all NCAA Softball rosters for a given seaosn and division}
\usage{
ncaa_softball_rosters(season, division)
}
\arguments{
\item{season}{2021-2023}

\item{division}{"D1", "D2", or "D3}
}
\value{
dataframe of every player on every team in a given season/division
}
\description{
Get all NCAA Softball rosters for a given seaosn and division
}
\examples{
try(ncaa_softball_rosters(2023, "D1"))
}
