% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_softball_scoreboard.R
\name{ncaa_softball_scoreboard}
\alias{ncaa_softball_scoreboard}
\title{Get all NCAA softball scores for a given day}
\usage{
ncaa_softball_scoreboard(date, division = "D1")
}
\arguments{
\item{date}{"YYYY-MM-DD"}

\item{division}{"D1", "D2", or "D3"}
}
\value{
data frame of date, team names and their scores
}
\description{
Supports 2017-2023 for D1 and only 2023 for D2 and D3
}
\examples{
date = "2022-03-05"
try(ncaa_softball_scoreboard(date))
}
\author{
Tyson King
}
