% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_softball_teams.R
\name{ncaa_softball_teams}
\alias{ncaa_softball_teams}
\title{Get all NCAA softball team names and ids}
\usage{
ncaa_softball_teams(season, division = "D1")
}
\arguments{
\item{season}{2021, 2022 2023, or 2024}

\item{division}{"D1", "D2", or "D3}
}
\value{
data frame of date, team names and their scores
}
\description{
2021-2023 for D1, only 2023 for D2 and D3
}
\examples{
try(ncaa_softball_teams(2024))
}
\author{
Tyson King
}
