% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_player_box.R
\name{get_ncaa_player_box}
\alias{get_ncaa_player_box}
\title{Get all box scores for a given game id}
\usage{
get_ncaa_player_box(game_id)
}
\arguments{
\item{team_id}{get game ids from get_ncaa_scoreboard function}
}
\value{
named list with three elements, "Hitting", "Pitching" and "Fielding".
Each contains a dataframe of the box scores for that game
}
\description{
Get all box scores for a given game id
}
\examples{
try(get_ncaa_player_box(2377042))
}
