% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_scoreboard.R
\name{get_ncaa_scoreboard}
\alias{get_ncaa_scoreboard}
\title{Get all NCAA softball scores for a given day}
\usage{
get_ncaa_scoreboard(date)
}
\arguments{
\item{date}{"YYYY-MM-DD"}
}
\value{
data frame of date, team names and their scores
}
\description{
for now only supports 2022 and 2023 seasons, will update further later
}
\examples{
date = "2022-03-05"
try(get_ncaa_scoreboard(date))
}
\author{
Tyson King
}
