% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_season_scoreboard.R
\name{get_ncaa_season_scoreboard}
\alias{get_ncaa_season_scoreboard}
\title{Get all NCAA softball scores for a given season}
\usage{
get_ncaa_season_scoreboard(season)
}
\arguments{
\item{season}{YYYY}
}
\value{
data frame of date, team names and their scores
}
\description{
for now only supports 2022 and 2023 seasons, will update further later
}
\examples{
season = "2023"
try(get_ncaa_season_scoreboard(season))
}
