% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_espn_scoreboard.R
\name{load_espn_scoreboard}
\alias{load_espn_scoreboard}
\title{Load ESPN Scoreboard}
\usage{
load_espn_scoreboard(seasons)
}
\arguments{
\item{seasons}{vector of seasons (YYYY:YYYY)}
}
\value{
dataframe of all games on ESPN website from selected seasons
}
\description{
Load ESPN Scoreboard
}
\examples{
try(load_espn_scoreboard(2020:2023))
}
