% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_scoreboard.R
\name{load_ncaa_scoreboard}
\alias{load_ncaa_scoreboard}
\title{Load NCAA Scoreboard}
\usage{
load_ncaa_scoreboard(season)
}
\arguments{
\item{season}{season (YYYY)}
}
\value{
dataframe of all games from NCAA website from selected season
}
\description{
ONLY HAS 2023 DATA FOR NOW
}
\examples{
try(load_ncaa_scoreboard(2023))
}
