% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_playerbox.R
\name{load_ncaa_softball_playerbox}
\alias{load_ncaa_softball_playerbox}
\title{Load NCAA Player Box Scores}
\usage{
load_ncaa_softball_playerbox(season = 2023, category, division = "D1")
}
\arguments{
\item{season}{YYYY (2023)}

\item{category}{Hitting, Pitching, or Fielding}
}
\value{
dataframe of all box scores from NCAA website from selected season
}
\description{
Only supports 2023 for now (only D1 for fielding)
}
\examples{
try(load_ncaa_softball_playerbox(2023))
}
