% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_playerbox.R
\name{load_ncaa_softball_playerbox}
\alias{load_ncaa_softball_playerbox}
\title{Load NCAA Player Box Scores}
\usage{
load_ncaa_softball_playerbox(season = 2023, category, division = "D1")
}
\arguments{
\item{season}{YYYY can also use vector of years}

\item{category}{Hitting, Pitching, or Fielding}

\item{division}{D1, D2, or D3}
}
\value{
dataframe of all box scores from NCAA website from selected season
}
\description{
2021-2023 for D1 hitting and pitching, only 2023 for everything else
}
\examples{
try(load_ncaa_softball_playerbox(2021:2023, category = "Hitting"))
}
