% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_softball_scoreboard.R
\name{load_ncaa_softball_scoreboard}
\alias{load_ncaa_softball_scoreboard}
\title{Load NCAA Scoreboard}
\usage{
load_ncaa_softball_scoreboard(season, division = "D1")
}
\arguments{
\item{season}{YYYY format, pass season or vector of seasons}

\item{division}{"D1", "D2", or "D3"}
}
\value{
dataframe of all games from NCAA website from selected season
}
\description{
Supports 2016-2023 for all divisions
}
\examples{
try(load_ncaa_softball_scoreboard(2022:2024))
}
