% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_softball_teams.R
\name{ncaa_softball_teams}
\alias{ncaa_softball_teams}
\title{Get all NCAA softball team names and ids}
\usage{
ncaa_softball_teams(season)
}
\arguments{
\item{season}{2021, 2022 or 2023}
}
\value{
data frame of date, team names and their scores
}
\description{
Only has division 1 teams for now
}
\examples{
try(ncaa_softball_teams(2023))
}
\author{
Tyson King
}
